requirements.txt
py-cord == 2.5.0
python-dotenv == 1.*
PyYAML == 6.0.1